Move mask library to utilities
