+++ training/train.py
@@ -81,7 +81,7 @@ import torch.utils.data
 import torchvision.transforms as transforms
 from pathlib import Path
 
-import masks as masklib
+import utilities.masklib
 import stack_dataset
 from utilities.archive import ModelArchive
 from utilities.helpers import (gridsample_residual, save_chunk,
